arch - remove “Core” concept and simplify some things
